redirect user to home page if already logged in and on landing page
